# Conflicts:
#	conftest.py
#	page_element/house/detail.yaml

+++ TestCase/test_sale/test_house/test_upload_agreement.py
@@ -16,7 +16,6 @@ from utils.jsonutil import get_value
 from page_object.main.topviewpage import MainTopViewPage
 from page_object.main.leftviewpage import MainLeftViewPage
 from page_object.main.rightviewpage import MainRightViewPage
-from page_object.main.upviewpage import MainUpViewPage
 from page_object.main.certificateexaminepage import CertificateExaminePage
 from page_object.house.tablepage import HouseTablePage
 from page_object.house.detailpage import HouseDetailPage
@@ -42,9 +41,7 @@ class TestUploadAgreement(object):
         global house_code
 
         main_leftview = MainLeftViewPage(web_driver)
-        main_upview = MainUpViewPage(web_driver)
         house_table = HouseTablePage(web_driver)
-        house_detail = HouseDetailPage(web_driver)
 
         main_leftview.change_role('经纪人')
         house_code = house_table.get_house_code_by_db(flag='买卖')
@@ -60,28 +57,34 @@ class TestUploadAgreement(object):
         main_leftview = MainLeftViewPage(web_driver)
         main_rightview = MainRightViewPage(web_driver)
         main_topview = MainTopViewPage(web_driver)
-        main_upview = MainUpViewPage(web_driver)
         house_table = HouseTablePage(web_driver)
         house_detail = HouseDetailPage(web_driver)
         agreement_list = AgreementListPage(web_driver)
         certificate_examine = CertificateExaminePage(web_driver)
 
         main_leftview.click_agreement_list_label()  # 获取协议编号
-        agreement_list.input_agreement_name_search('一般委托书')
-        agreement_list.click_query_button()
-        agreement_list.click_download_button_by_row(1)
-        written_entrustment_agreement_number = agreement_list.get_written_entrustment_agreement_number()
-        self.written_entrustment_agreement['委托协议编号'] = written_entrustment_agreement_number
-        agreement_list.input_agreement_name_search('钥匙托管协议')
-        agreement_list.click_query_button()
-        agreement_list.click_download_button_by_row(1)
-        key_entrustment_certificate_number = agreement_list.get_key_entrustment_certificate_number()
-        self.key_entrustment_certificate['协议编号'] = key_entrustment_certificate_number
-        agreement_list.input_agreement_name_search('房屋出售委托协议VIP版')
-        agreement_list.click_query_button()
-        agreement_list.click_download_button_by_row(1)
-        vip_service_entrustment_agreement_number = agreement_list.get_vip_service_entrustment_agreement_number()
-        self.vip_service_entrustment_agreement['委托协议编号'] = vip_service_entrustment_agreement_number
+        if ini.environment == 'sz':
+            agreement_list.input_agreement_name_search('一般委托书')
+            agreement_list.click_query_button()
+            agreement_list.click_download_button_by_row(1)
+            written_entrustment_agreement_number = agreement_list.get_written_entrustment_agreement_number()
+            self.written_entrustment_agreement['委托协议编号'] = written_entrustment_agreement_number
+            agreement_list.input_agreement_name_search('钥匙托管协议')
+            agreement_list.click_query_button()
+            agreement_list.click_download_button_by_row(1)
+            key_entrustment_certificate_number = agreement_list.get_key_entrustment_certificate_number()
+            self.key_entrustment_certificate['协议编号'] = key_entrustment_certificate_number
+            agreement_list.input_agreement_name_search('房屋出售委托协议VIP版')
+            agreement_list.click_query_button()
+            agreement_list.click_download_button_by_row(1)
+            vip_service_entrustment_agreement_number = agreement_list.get_vip_service_entrustment_agreement_number()
+            self.vip_service_entrustment_agreement['委托协议编号'] = vip_service_entrustment_agreement_number
+        if ini.environment == 'wx':
+            agreement_list.input_agreement_name_search('限时委托代理销售协议')
+            agreement_list.click_query_button()
+            agreement_list.click_download_button_by_row(1)
+            written_entrustment_agreement_number = agreement_list.get_written_entrustment_agreement_number()
+            self.written_entrustment_agreement['委托协议编号'] = written_entrustment_agreement_number
 
         main_leftview.click_all_house_label()  # 进入房源详情
         house_table.click_sale_tab()
@@ -90,7 +93,7 @@ class TestUploadAgreement(object):
         house_table.clear_filter('买卖')
         house_table.input_house_code_search(house_code)
         house_table.click_search_button()
-        house_table.go_house_detail_by_row()
+        house_table.go_house_detail_by_row(1)
         house_detail.expand_certificates_info()  # 上传协议
         if house_detail.check_upload_written_entrustment_agreement() != '未上传':
             house_detail.delete_written_entrustment_agreement()
@@ -100,24 +103,25 @@ class TestUploadAgreement(object):
         house_detail.upload_written_entrustment_agreement(self.written_entrustment_agreement)
         assert main_topview.find_notification_content() == '上传成功'
         log.info('书面委托协议已上传')
-        house_detail.expand_certificates_info()
-        if house_detail.check_upload_key_entrustment_certificate() != '未上传':
-            house_detail.delete_key_entrustment_certificate()
-            assert main_topview.find_notification_content() == '操作成功'
-            house_detail.page_refresh()
+        if ini.environment == 'sz':
             house_detail.expand_certificates_info()
-        house_detail.upload_key_entrustment_certificate(self.key_entrustment_certificate)
-        assert main_topview.find_notification_content() == '上传成功'
-        log.info('钥匙委托凭证已上传')
-        house_detail.expand_certificates_info()
-        if house_detail.check_upload_vip_service_entrustment_agreement() != '未上传':
-            house_detail.delete_vip_service_entrustment_agreement()
-            assert main_topview.find_notification_content() == '操作成功'
-            house_detail.page_refresh()
+            if house_detail.check_upload_key_entrustment_certificate() != '未上传':
+                house_detail.delete_key_entrustment_certificate()
+                assert main_topview.find_notification_content() == '操作成功'
+                house_detail.page_refresh()
+                house_detail.expand_certificates_info()
+            house_detail.upload_key_entrustment_certificate(self.key_entrustment_certificate)
+            assert main_topview.find_notification_content() == '上传成功'
+            log.info('钥匙委托凭证已上传')
             house_detail.expand_certificates_info()
-        house_detail.upload_vip_service_entrustment_agreement(self.vip_service_entrustment_agreement)
-        assert main_topview.find_notification_content() == '上传成功'
-        log.info('VIP服务委托协议已上传')
+            if house_detail.check_upload_vip_service_entrustment_agreement() != '未上传':
+                house_detail.delete_vip_service_entrustment_agreement()
+                assert main_topview.find_notification_content() == '操作成功'
+                house_detail.page_refresh()
+                house_detail.expand_certificates_info()
+            house_detail.upload_vip_service_entrustment_agreement(self.vip_service_entrustment_agreement)
+            assert main_topview.find_notification_content() == '上传成功'
+            log.info('VIP服务委托协议已上传')
         house_detail.expand_certificates_info()
         if house_detail.check_upload_deed_tax_invoice() != '未上传':
             house_detail.delete_deed_tax_invoice()
@@ -158,7 +162,7 @@ class TestUploadAgreement(object):
         main_rightview.click_certificate_examine()
         assert certificate_examine.get_table_count() > 0
         certificate_examine.pass_written_entrustment_agreement_examine(house_code)
-        certificate_examine.pass_key_entrustment_certificate_examine(house_code)
-        certificate_examine.pass_vip_service_entrustment_agreement_examine(house_code)
+        if ini.environment == 'sz':
+            certificate_examine.pass_key_entrustment_certificate_examine(house_code)
+            certificate_examine.pass_vip_service_entrustment_agreement_examine(house_code)
         certificate_examine.pass_property_ownership_certificate_examine(house_code)
-        main_upview.clear_all_title()
